UHF-5265: Better role map logic
